move config command to archiver.config
