+++ setup.cfg
@@ -118,6 +118,7 @@ per_file_ignores =
     src/borg/archive.py:E122,E125,E127,E402,E501,F401,F405,W504
     src/borg/archiver/__init__.py:E402,E501,E722,E741,F405
     src/borg/archiver/benchmarks.py:F405
+    src/borg/archiver/config.py:F405,E722
     src/borg/archiver/common.py:E501,F405
     src/borg/archiver/debug.py:F405
     src/borg/archiver/help.py:E501,F405
